Add files via upload
Simple shot clock code for Arduino. 

Runs 2 SparkFun 7-segment displays (really 8 segment, as they include a decimal) as well as a smaller QWIIC 4 number display for the user-facing side.

Expects 3 inputs: a "running" switch, a 25 or 15 second switch, and a reset button.

The timeLeft decrements in ~100ms jumps. When timeLeft > 10 seconds, both digits are used to show the number of seconds remaining. When the timeLeft drops below 10 seconds, the display turns on the middle decimal LED and displays <seconds>.<10ths of a second>.